asm: remove some already-converted AVX instructions
